# backend/src/main/java/ru/skfu/moviecollection/control/dto/AdminStatsDto.java
package ru.skfu.moviecollection.control.dto;

public record AdminStatsDto(
        long usersCount,
        long moviesCount,
        long collectionItemsCount,
        long openComplaintsCount
) {
}